+++ bridge-ui/src/components/bridge/modals/ManualClaimModal.tsx
@@ -10,7 +10,7 @@ type ManualClaimModalProps = {
 const ManualClaimModal: React.FC<ManualClaimModalProps> = ({ handleNoClose, handleYesClose }) => {
   return (
     <div className="flex flex-col items-center justify-center gap-8 px-8">
-      <h2 className="text-2xl text-[#E5E5E5]">Manual claim on destination</h2>
+      <h2 className="text-2xl">Manual claim on destination</h2>
       <div className="space-y-5 text-center">
         <div className="text-sm">
           Activating Manual claim means you will need to claim the message on the destination chain with ETH with a
@@ -26,10 +26,10 @@ const ManualClaimModal: React.FC<ManualClaimModalProps> = ({ handleNoClose, hand
             href="https://docs.linea.build/developers/guides/bridge/how-to-bridge-eth"
             rel="noopener noreferrer"
             target="_blank"
-            className="hover:border-b-1 border-b-1 btn btn-ghost btn-sm rounded-none border-b-primary p-0 font-normal text-[#E5E5E5] hover:border-b-primary hover:bg-transparent"
+            className="hover:border-b-1 border-b-1 btn btn-ghost btn-sm rounded-none border-b-primary p-0 font-normal hover:border-b-primary hover:bg-transparent"
           >
             More Info
-            <MdCallMade color="white" />
+            <MdCallMade />
           </Link>
         </div>
       </div>