+++ contracts/hardhat.config.ts
@@ -1,4 +1,5 @@
 import "@nomicfoundation/hardhat-toolbox";
+import "@nomicfoundation/hardhat-foundry";
 import "@openzeppelin/hardhat-upgrades";
 import "@nomicfoundation/hardhat-foundry";
 import * as dotenv from "dotenv";
@@ -12,7 +13,10 @@ import "./scripts/operational/grantContractRolesTask";
 import "./scripts/operational/renounceContractRolesTask";
 import "./scripts/operational/setRateLimitTask";
 import "./scripts/operational/setVerifierAddressTask";
-import "./scripts/operational/transferOwnershipAndSetRemoteTokenBridgeTask";
+import "./scripts/operational/setRemoteTokenBridgeTask";
+import "./scripts/operational/setMessageServiceOnTokenBridgeTask";
+
+import "solidity-docgen";
 
 dotenv.config();
 
@@ -149,6 +153,11 @@ const config: HardhatUserConfig = {
         },
       },
     ],
+  },
+  docgen: {
+    exclude: ["token", "test-contracts", "proxies", "tools", "interfaces/tools", "tokenBridge/mocks", "verifiers"],
+    pages: "files",
+    outputDir: "docs/api/",
   },
 };