+++ prover/backend/aggregation/prove.go
@@ -5,9 +5,9 @@ import (
 	"math"
 	"path/filepath"
 
-	frBls12377 "github.com/consensys/gnark-crypto/ecc/bls12-377/fr"
 	"github.com/consensys/gnark/backend/witness"
 	"github.com/consensys/gnark/frontend"
+	emPlonk "github.com/consensys/gnark/std/recursion/plonk"
 	pi_interconnection "github.com/consensys/linea-monorepo/prover/circuits/pi-interconnection"
 	public_input "github.com/consensys/linea-monorepo/prover/public-input"
 
@@ -67,7 +67,37 @@ func makeProof(
 	return circuits.SerializeProofSolidityBn254(proofBn254), nil
 }
 
+func (cf CollectedFields) AggregationPublicInput(cfg *config.Config) public_input.Aggregation {
+	return public_input.Aggregation{
+		FinalShnarf:                             cf.FinalShnarf,
+		ParentAggregationFinalShnarf:            cf.ParentAggregationFinalShnarf,
+		ParentStateRootHash:                     cf.ParentStateRootHash,
+		ParentAggregationLastBlockTimestamp:     cf.ParentAggregationLastBlockTimestamp,
+		FinalTimestamp:                          cf.FinalTimestamp,
+		LastFinalizedBlockNumber:                cf.LastFinalizedBlockNumber,
+		FinalBlockNumber:                        cf.FinalBlockNumber,
+		LastFinalizedL1RollingHash:              cf.LastFinalizedL1RollingHash,
+		L1RollingHash:                           cf.L1RollingHash,
+		LastFinalizedL1RollingHashMessageNumber: cf.LastFinalizedL1RollingHashMessageNumber,
+		L1RollingHashMessageNumber:              cf.L1RollingHashMessageNumber,
+		L2MsgRootHashes:                         cf.L2MsgRootHashes,
+		L2MsgMerkleTreeDepth:                    utils.ToInt(cf.L2MsgTreeDepth),
+		ChainID:                                 uint64(cfg.Layer2.ChainID),
+		L2MessageServiceAddr:                    types.EthAddress(cfg.Layer2.MsgSvcContract),
+	}
+}
+
 func makePiProof(cfg *config.Config, cf *CollectedFields) (plonk.Proof, witness.Witness, error) {
+
+	var setup circuits.Setup
+	setupErr := make(chan error, 1)
+
+	go func() {
+		var err error
+		setup, err = circuits.LoadSetup(cfg, circuits.PublicInputInterconnectionCircuitID)
+		setupErr <- err
+		close(setupErr)
+	}()
 
 	c, err := pi_interconnection.Compile(cfg.PublicInputInterconnection, pi_interconnection.WizardCompilationParameters()...)
 	if err != nil {
@@ -77,29 +107,10 @@ func makePiProof(cfg *config.Config, cf *CollectedFields) (plonk.Proof, witness.
 	assignment, err := c.Assign(pi_interconnection.Request{
 		Decompressions: cf.DecompressionPI,
 		Executions:     cf.ExecutionPI,
-		Aggregation: public_input.Aggregation{
-			FinalShnarf:                             cf.FinalShnarf,
-			ParentAggregationFinalShnarf:            cf.ParentAggregationFinalShnarf,
-			ParentStateRootHash:                     cf.ParentStateRootHash,
-			ParentAggregationLastBlockTimestamp:     cf.ParentAggregationLastBlockTimestamp,
-			FinalTimestamp:                          cf.FinalTimestamp,
-			LastFinalizedBlockNumber:                cf.LastFinalizedBlockNumber,
-			FinalBlockNumber:                        cf.FinalBlockNumber,
-			LastFinalizedL1RollingHash:              cf.LastFinalizedL1RollingHash,
-			L1RollingHash:                           cf.L1RollingHash,
-			LastFinalizedL1RollingHashMessageNumber: cf.LastFinalizedL1RollingHashMessageNumber,
-			L1RollingHashMessageNumber:              cf.L1RollingHashMessageNumber,
-			L2MsgRootHashes:                         cf.L2MsgRootHashes,
-			L2MsgMerkleTreeDepth:                    utils.ToInt(cf.L2MsgTreeDepth),
-		},
+		Aggregation:    cf.AggregationPublicInput(cfg),
 	})
 	if err != nil {
 		return nil, nil, fmt.Errorf("could not assign the public input circuit: %w", err)
-	}
-
-	setup, err := circuits.LoadSetup(cfg, circuits.PublicInputInterconnectionCircuitID)
-	if err != nil {
-		return nil, nil, fmt.Errorf("could not load the setup: %w", err)
 	}
 
 	w, err := frontend.NewWitness(&assignment, ecc.BLS12_377.ScalarField(), frontend.PublicOnly()) // TODO @Tabaie make ProveCheck return witness instead of extracting this twice
@@ -107,7 +118,14 @@ func makePiProof(cfg *config.Config, cf *CollectedFields) (plonk.Proof, witness.
 		return nil, nil, fmt.Errorf("could not extract interconnection circuit public witness: %w", err)
 	}
 
-	proof, err := circuits.ProveCheck(&setup, &assignment)
+	if err = <-setupErr; err != nil { // wait for setup to load and check for errors
+		return nil, nil, fmt.Errorf("could not load the setup: %w", err)
+	}
+
+	proverOpts := emPlonk.GetNativeProverOptions(ecc.BW6_761.ScalarField(), setup.Circuit.Field())
+	verifierOpts := emPlonk.GetNativeVerifierOptions(ecc.BW6_761.ScalarField(), setup.Circuit.Field())
+
+	proof, err := circuits.ProveCheck(&setup, &assignment, proverOpts, verifierOpts)
 
 	return proof, w, err
 }
@@ -218,13 +236,9 @@ func makeBw6Proof(
 	// single field element.
 
 	var (
-		piBW6      frBW6.Element
-		piBls12377 frBls12377.Element
+		piBW6 frBW6.Element
 	)
 	if _, err = piBW6.SetString(publicInput); err != nil {
-		return nil, 0, fmt.Errorf("could not parse the public input: %w", err)
-	}
-	if _, err = piBls12377.SetString(publicInput); err != nil {
 		return nil, 0, fmt.Errorf("could not parse the public input: %w", err)
 	}