+++ prover/circuits/blobdecompression/large-tests/maximize_blob_size/main.go
@@ -1,19 +1,19 @@
 package main
 
 import (
+	"flag"
 	"fmt"
-
 	"github.com/consensys/gnark-crypto/ecc"
 	"github.com/consensys/gnark/frontend"
 	"github.com/consensys/gnark/frontend/cs/scs"
 	v1 "github.com/consensys/linea-monorepo/prover/circuits/blobdecompression/v1"
 	blob "github.com/consensys/linea-monorepo/prover/lib/compressor/blob/v1"
+	"runtime"
 )
 
 const maxNbConstraints = 1 << 27
 
 func nbConstraints(blobSize int) int {
-
 	fmt.Printf("*********************\nfor blob of size %dB or %.2fKB:\n", blobSize, float32(blobSize)/1024)
 	c := v1.Circuit{
 		BlobBytes:             make([]frontend.Variable, 32*4096),
@@ -21,6 +21,7 @@ func nbConstraints(blobSize int) int {
 		MaxBlobPayloadNbBytes: blobSize,
 		UseGkrMiMC:            true,
 	}
+	runtime.GC()
 	if cs, err := frontend.Compile(ecc.BLS12_377.ScalarField(), scs.NewBuilder, &c, frontend.WithCapacity(maxNbConstraints*6/5)); err != nil {
 		panic(err)
 	} else {
@@ -37,27 +38,32 @@ func nbConstraints(blobSize int) int {
 	}
 }
 
+var (
+	flagCrawlStep = flag.Int("step", 1000, "the crawl step") // TODO @Tabaie fix mixed metaphor
+	flagStart     = flag.Int("start", blob.MaxUncompressedBytes, "initial size in bytes")
+)
+
 func main() {
 
-	crawlStep := 1000 // TODO bad name; mixed metaphor
-	v := nbConstraints(blob.MaxUncompressedBytes)
-	a, b := blob.MaxUncompressedBytes, blob.MaxUncompressedBytes
+	flag.Parse()
+	v := nbConstraints(*flagStart)
+	a, b := *flagStart, *flagStart
 
 	if v > maxNbConstraints {
 		fmt.Println("crawling downward")
 		for v > maxNbConstraints {
 			b = a
-			a = max(a-crawlStep, 0)
+			a = max(a-*flagCrawlStep, 0)
 			v = nbConstraints(a)
-			crawlStep *= 2
+			*flagCrawlStep *= 2
 		}
 	} else if v < maxNbConstraints {
 		fmt.Println("crawling upward")
 		for v < maxNbConstraints {
 			a = b
-			b += crawlStep
+			b += *flagCrawlStep
 			v = nbConstraints(b)
-			crawlStep *= 2
+			*flagCrawlStep *= 2
 		}
 	}
 	if v == maxNbConstraints {