+++ prover/utils/test_utils/test_utils.go
@@ -12,9 +12,11 @@ import (
 	"math"
 	"os"
 	"reflect"
-	"strconv"
 	"strings"
 	"testing"
+
+	"github.com/consensys/gnark/frontend"
+	snarkHash "github.com/consensys/gnark/std/hash"
 
 	"github.com/stretchr/testify/require"
 )
@@ -49,98 +51,10 @@ func RandIntSliceN(length, n int) []int {
 	return res
 }
 
-type BytesEqualError struct {
-	Index int
-	error string
-}
-
-func (e *BytesEqualError) Error() string {
-	return e.error
-}
-
 func LoadJson(t *testing.T, path string, v any) {
 	in, err := os.Open(path)
 	require.NoError(t, err)
 	require.NoError(t, json.NewDecoder(in).Decode(v))
-}
-
-// BytesEqual between byte slices a,b
-// a readable error message would show in case of inequality
-// TODO error options: block size, check forwards or backwards etc
-func BytesEqual(expected, actual []byte) error {
-	l := min(len(expected), len(actual))
-
-	failure := 0
-	for failure < l {
-		if expected[failure] != actual[failure] {
-			break
-		}
-		failure++
-	}
-
-	if len(expected) == len(actual) {
-		return nil
-	}
-
-	// there is a mismatch
-	var sb strings.Builder
-
-	const (
-		radius    = 40
-		blockSize = 32
-	)
-
-	printCentered := func(b []byte) {
-
-		for i := max(failure-radius, 0); i <= failure+radius; i++ {
-			if i%blockSize == 0 && i != failure-radius {
-				sb.WriteString("  ")
-			}
-			if i >= 0 && i < len(b) {
-				sb.WriteString(hex.EncodeToString([]byte{b[i]})) // inefficient, but this whole error printing sub-procedure will not be run more than once
-			} else {
-				sb.WriteString("  ")
-			}
-		}
-	}
-
-	sb.WriteString(fmt.Sprintf("mismatch starting at byte %d\n", failure))
-
-	sb.WriteString("expected: ")
-	printCentered(expected)
-	sb.WriteString("\n")
-
-	sb.WriteString("actual:   ")
-	printCentered(actual)
-	sb.WriteString("\n")
-
-	sb.WriteString("          ")
-	for i := max(failure-radius, 0); i <= failure+radius; {
-		if i%blockSize == 0 && i != failure-radius {
-			s := strconv.Itoa(i)
-			sb.WriteString("  ")
-			sb.WriteString(s)
-			i += len(s) / 2
-			if len(s)%2 != 0 {
-				sb.WriteString(" ")
-				i++
-			}
-		} else {
-			if i == failure {
-				sb.WriteString("^^")
-			} else {
-				sb.WriteString("  ")
-			}
-			i++
-		}
-	}
-
-	sb.WriteString("\n")
-
-	return &BytesEqualError{
-		Index: failure,
-		error: sb.String(),
-	}
 }
 
 func SlicesEqual[T any](expected, actual []T) error {
@@ -186,6 +100,7 @@ func (w *WriterHash) Reset() {
 	if _, err := w.w.Write([]byte{255, 255}); err != nil {
 		panic(err)
 	}
+	w.h.Reset()
 }
 
 func (w *WriterHash) Size() int {
@@ -220,22 +135,10 @@ type ReaderHash struct {
 }
 
 func (r *ReaderHash) Write(p []byte) (n int, err error) {
-	if len(p) >= 65535 {
-		panic("ReaderHash.Write: too large")
+	if rd := hashReadWrite(r.r); !bytes.Equal(rd, p) {
+		panic(fmt.Errorf("ReaderHash.Write: expected %x, encountered %x", rd, p))
 	}
 
-	var ls [2]byte
-	if _, err = r.r.Read(ls[:]); err != nil {
-		panic(err)
-	}
-
-	buf := make([]byte, int(ls[0])*256+int(ls[1]))
-	if _, err = r.r.Read(buf); err != nil {
-		panic(err)
-	}
-	if !bytes.Equal(buf, p) {
-		panic(fmt.Errorf("ReaderHash.Write: mismatch %x≠%x", buf, p))
-	}
 	return r.h.Write(p)
 }
 
@@ -247,14 +150,36 @@ func (r *ReaderHash) Sum(b []byte) []byte {
 }
 
 func (r *ReaderHash) Reset() {
+	hashReadReset(r.r)
+	r.h.Reset()
+}
+
+func hashReadWrite(r io.Reader) []byte {
+
 	var ls [2]byte
-	if _, err := r.r.Read(ls[:]); err != nil {
+	if _, err := r.Read(ls[:]); err != nil {
+		panic(err)
+	}
+
+	buf := make([]byte, int(ls[0])*256+int(ls[1]))
+	if len(buf) == 65535 {
+		panic("ReaderHash.Write: Reset expected")
+	}
+	if _, err := r.Read(buf); err != nil {
+		panic(err)
+	}
+
+	return buf
+}
+
+func hashReadReset(r io.Reader) {
+	var ls [2]byte
+	if _, err := r.Read(ls[:]); err != nil {
 		panic(err)
 	}
 	if ls[0] != 255 || ls[1] != 255 {
 		panic(fmt.Errorf("ReaderHash.Reset: unexpected %x", ls))
 	}
-	r.h.Reset()
 }
 
 func (r *ReaderHash) Size() int {
@@ -280,4 +205,145 @@ func (r *ReaderHash) CloseFile() {
 	if err := r.r.(*os.File).Close(); err != nil {
 		panic(err)
 	}
+}
+
+// ReaderHashSnark is a wrapper around a FieldHasher that matches all writes with its input stream.
+type ReaderHashSnark struct {
+	h   snarkHash.FieldHasher
+	r   io.Reader
+	api frontend.API
+}
+
+func NewReaderHashSnarkFromFile(api frontend.API, h snarkHash.FieldHasher, path string) snarkHash.FieldHasher {
+	r, err := os.Open(path)
+	if err != nil {
+		panic(err)
+	}
+	return &ReaderHashSnark{
+		h:   h,
+		r:   r,
+		api: api,
+	}
+}
+
+func (r *ReaderHashSnark) Sum() frontend.Variable {
+	return r.h.Sum()
+}
+
+func (r *ReaderHashSnark) Write(data ...frontend.Variable) {
+	r.h.Write(data...)
+
+	for i := 0; i < len(data); {
+		buf := hashReadWrite(r.r)
+		for len(buf) != 0 {
+			n := min(len(buf), (r.api.Compiler().FieldBitLen()+7)/8)
+			r.api.AssertIsEqual(data[i], buf[:n])
+			buf = buf[n:]
+			i++
+		}
+	}
+}
+
+func (r *ReaderHashSnark) Reset() {
+	hashReadReset(r.r)
+	r.h.Reset()
+}
+
+func (r *ReaderHashSnark) CloseFile() {
+	if err := r.r.(*os.File).Close(); err != nil {
+		panic(err)
+	}
+}
+
+// PrettyPrintHashes reads a file of hashes and prints them in a human-readable format.
+// if out is nil, it will print to os.Stdout
+func PrettyPrintHashes(in string, out io.Writer) {
+	const printIndexes = false
+	nbResets := 0
+
+	v := make([]any, 0)
+	f, err := os.Open(in)
+	if err != nil {
+		panic(err)
+	}
+	var (
+		length [2]byte
+		buf    []byte
+	)
+
+	var i int
+	for _, err = f.Read(length[:]); err == nil; _, err = f.Read(length[:]) {
+		l := int(length[0])*256 + int(length[1])
+		if l == 65535 { // a reset
+			v = append(v, fmt.Sprintf("RESET #%d", nbResets))
+			nbResets++
+			i = 0
+			continue
+		}
+		if l > len(buf) {
+			buf = make([]byte, l)
+		}
+
+		if _, err = f.Read(buf[:l]); err != nil {
+			break
+		}
+
+		prettyBuf := spaceOutFromRight(hex.EncodeToString(buf[:l]))
+		if printIndexes {
+			v = append(v, fmt.Sprintf("%d: 0x%s", i, prettyBuf))
+		} else {
+			v = append(v, "0x"+prettyBuf)
+		}
+
+		i++
+	}
+	if err != io.EOF {
+		panic(err)
+	}
+
+	if out == nil {
+		out = os.Stdout
+	}
+	encoder := json.NewEncoder(out)
+	encoder.SetIndent("", "  ")
+	if err = encoder.Encode(v); err != nil {
+		panic(err)
+	}
+}
+
+func PrettyPrintHashesToFile(in, out string) {
+	var t FakeTestingT
+	f, err := os.OpenFile(out, os.O_CREATE|os.O_WRONLY, 0600)
+	require.NoError(t, err)
+	PrettyPrintHashes(in, f)
+	require.NoError(t, err)
+}
+
+func spaceOutFromRight(s string) string {
+	n := len(s) + (len(s)+15)/16 - 1
+	if n < 0 {
+		return ""
+	}
+	var bb strings.Builder
+	bb.Grow(n)
+	remainder := len(s) % 16
+	first := true
+	if remainder != 0 {
+		bb.WriteString(s[:remainder])
+		s = s[remainder:]
+		first = false
+	}
+	for len(s) > 0 {
+		if !first {
+			bb.WriteByte(' ')
+		}
+		bb.WriteString(s[:16])
+		s = s[16:]
+		first = false
+	}
+
+	if bb.Len() != n {
+		panic("incorrect size estimation")
+	}
+	return bb.String()
 }