+++ transaction-exclusion-api/app/src/test/kotlin/net/consensys/linea/transactionexclusion/app/api/ArgumentParserTest.kt
@@ -6,7 +6,8 @@ import net.consensys.decodeHex
 import net.consensys.encodeHex
 import net.consensys.linea.transactionexclusion.ModuleOverflow
 import net.consensys.linea.transactionexclusion.RejectedTransaction
-import net.consensys.linea.transactionexclusion.TransactionInfo
+import net.consensys.linea.transactionexclusion.test.defaultRejectedTransaction
+import net.consensys.linea.transactionexclusion.test.rejectedContractDeploymentTransaction
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
@@ -15,9 +16,7 @@ import kotlin.random.Random
 class ArgumentParserTest {
   @Test
   fun getTransactionRLPInRawBytes_should_return_correct_byte_array() {
-    val transactionRLPInHexStr =
-      "0x02f8388204d2648203e88203e88203e8941195cf65f83b3a5768f3c4" +
-        "96d3a05ad6412c64b38203e88c666d93e9cc5f73748162cea9c0017b8201c8"
+    val transactionRLPInHexStr = defaultRejectedTransaction.transactionRLP.encodeHex()
     Assertions.assertTrue(
       ArgumentParser.getTransactionRLPInRawBytes(transactionRLPInHexStr)
         .contentEquals(transactionRLPInHexStr.decodeHex())
@@ -76,19 +75,19 @@ class ArgumentParserTest {
 
   @Test
   fun getTransactionInfoFromRLP_should_return_correct_transactionInfo() {
-    val transactionRLP =
-      (
-        "0x02f8388204d2648203e88203e88203e8941195cf65f83b3a5768f3c4" +
-          "96d3a05ad6412c64b38203e88c666d93e9cc5f73748162cea9c0017b8201c8"
-        ).decodeHex()
+    val transactionRLP = defaultRejectedTransaction.transactionRLP
     Assertions.assertEquals(
       ArgumentParser.getTransactionInfoFromRLP(transactionRLP),
-      TransactionInfo(
-        hash = "0x526e56101cf39c1e717cef9cedf6fdddb42684711abda35bae51136dbb350ad7".decodeHex(),
-        from = "0x4d144d7b9c96b26361d6ac74dd1d8267edca4fc2".decodeHex(),
-        to = "0x1195cf65f83b3a5768f3c496d3a05ad6412c64b3".decodeHex(),
-        nonce = 100UL
-      )
+      defaultRejectedTransaction.transactionInfo
+    )
+  }
+
+  @Test
+  fun getTransactionInfoFromRLP_should_return_correct_transactionInfo_for_contract_deployment_tx() {
+    val transactionRLP = rejectedContractDeploymentTransaction.transactionRLP
+    Assertions.assertEquals(
+      ArgumentParser.getTransactionInfoFromRLP(transactionRLP),
+      rejectedContractDeploymentTransaction.transactionInfo
     )
   }
 
@@ -217,15 +216,15 @@ class ArgumentParserTest {
   }
 
   @Test
-  fun getReasonMessage_should_throw_error_for_string_length_longer_than_256() {
-    // reason message string with more than 256 characters
+  fun getReasonMessage_should_throw_error_for_string_length_longer_than_1024() {
+    // reason message string with more than 1024 characters
     assertThrows<IllegalArgumentException> {
       ArgumentParser.getReasonMessage(
-        Random.Default.nextBytes(128).encodeHex(prefix = false) + "0"
+        Random.Default.nextBytes(512).encodeHex(prefix = false) + "0"
       )
     }.also { error ->
       Assertions.assertTrue(
-        error.message!!.contains("Reason message should not be more than 256 characters")
+        error.message!!.contains("Reason message should not be more than 1024 characters")
       )
     }
   }